server/requirements.txt
translators==6.0.4
flask==3.1.1
flask-cors==5.0.1
python-dotenv==1.1.0
